% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case-manipulation-utils.R
\name{add_uk}
\alias{add_uk}
\title{Adds a UK case count to a dataset usng national level data}
\usage{
add_uk(cases, min_uk)
}
\description{
Adds a UK case count to a dataset usng national level data
}
