% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate-derivative.R
\name{cd_prime_sources}
\alias{cd_prime_sources}
\title{cd_prime_sources
Calculate list of source files to process for a given target in a list of datasets}
\usage{
cd_prime_sources(datasets, target)
}
\arguments{
\item{datasets}{List of \code{DatasetLocation}}

\item{target}{String csv filename to collate}
}
\value{
List of files to collate
}
\description{
cd_prime_sources
Calculate list of source files to process for a given target in a list of datasets
}
