% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate-derivative.R
\name{cd_read_and_bind_sources}
\alias{cd_read_and_bind_sources}
\title{read_and_bind_sources}
\usage{
cd_read_and_bind_sources(sources)
}
\arguments{
\item{sources}{list of csv source files}
}
\value{
dataframe
}
\description{
read_and_bind_sources
}
