% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate-derivative.R
\name{collate_derivative}
\alias{collate_derivative}
\title{collate_derivative}
\usage{
collate_derivative(derivative, publish = TRUE)
}
\arguments{
\item{derivative}{\code{CollatedDerivative} object to calculate}

\item{publish}{Bool attempt publishing?}
}
\description{
collate_derivative
}
