% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collate_estimates}
\alias{collate_estimates}
\title{Collate estimates from different estimates in same sub-regional folder
DEPRECATED}
\usage{
collate_estimates(name, target = "rt")
}
\description{
Collate estimates from different estimates in same sub-regional folder
DEPRECATED
}
