% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_clean_cases}
\alias{generate_clean_cases}
\title{generate clean cases
produce a neat data set for a single region}
\usage{
generate_clean_cases(
  number_of_days = 90,
  regions = list("my_test_region"),
  peak_cases = 2000,
  days_since_peak = 0
)
}
\arguments{
\item{number_of_days}{\itemize{
\item how far back in time should we go
}}

\item{regions}{List Character region names}

\item{peak_cases}{What's the last value for number of new cases}

\item{days_since_peak}{Which day is the peak (0=today, 20 = 20 days ago)}
}
\description{
generate clean cases
produce a neat data set for a single region
}
