% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%in_ci\%}
\alias{\%in_ci\%}
\title{case insensitive version of the in function}
\usage{
x \%in_ci\% y
}
\arguments{
\item{x}{string}

\item{y}{string}
}
\value{
boolean
}
\description{
case insensitive version of the in function
}
