% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-region-updates.R
\name{loadStatsFile}
\alias{loadStatsFile}
\title{loadStatsFile}
\usage{
loadStatsFile(filename)
}
\arguments{
\item{filename}{String filename to load stats from (csv)}
}
\value{
data.frame of correctly formatted data - either loaded from the filename or blank if
filename is missing
}
\description{
loadStatsFile
}
