% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-region-updates.R
\name{loadStatusFile}
\alias{loadStatusFile}
\title{loadStatusFile
Clone of loadstatsfile but with different file structure}
\usage{
loadStatusFile(filename)
}
\arguments{
\item{filename}{String filename to load status from (csv)}
}
\value{
data.frame of correctly formatted data - either loaded from the filename or blank if
filename is missing
}
\description{
loadStatusFile
Clone of loadstatsfile but with different file structure
}
