% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_cludes}
\alias{parse_cludes}
\title{parse 'cludes}
\usage{
parse_cludes(cludes)
}
\arguments{
\item{cludes}{string of in/excludes}
}
\value{
data.frame of regions / subregions
}
\description{
parse 'cludes
}
