% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish-data.R
\name{publish_data}
\alias{publish_data}
\title{Publish a dataset to a dataverse}
\usage{
publish_data(dataset, files = TRUE, production_date = NA)
}
\arguments{
\item{dataset}{AbstractDataset to publish or CollatedDerivative (they both have the required interface)}

\item{files}{Boolean indicator as to if files should be uploaded. Helpful for test purposes}

\item{production_date}{Date specifying when the data is for. Gets used as the production date in the metadata, defaults to today.}
}
\description{
Publish a dataset to a dataverse
}
