% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-resummarise.R
\name{rrs_cli_interface}
\alias{rrs_cli_interface}
\title{rrs_cli_interface
Define the CLI interface and return the parsed arguments}
\usage{
rrs_cli_interface(args_string = NA)
}
\arguments{
\item{args_string}{String (optional) of command line flags to simulate CLI interface when running from
within another program / rstudio}
}
\value{
List of arguments
}
\description{
rrs_cli_interface
Define the CLI interface and return the parsed arguments
}
