% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-region-updates.R
\name{rru_cli_interface}
\alias{rru_cli_interface}
\title{rru_cli_interface
Define the CLI interface and return the parsed arguments}
\usage{
rru_cli_interface(args_string = NA)
}
\arguments{
\item{args_string}{String (optional) of command line flags to simulate CLI interface when running from
within another program / rstudio}
}
\value{
List of arguments
}
\description{
rru_cli_interface
Define the CLI interface and return the parsed arguments
}
