% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-region-updates.R
\name{rru_filter_datasets}
\alias{rru_filter_datasets}
\title{rru_filter_datasets
Slice out only the top level datasets we want to process}
\usage{
rru_filter_datasets(datasets, excludes, includes)
}
\arguments{
\item{datasets}{List of AbstractDataset}

\item{excludes}{List of strings, processed from the args}

\item{includes}{List of strings, processed from the args}
}
\value{
List of AbstractDataset or empty list
}
\description{
rru_filter_datasets
Slice out only the top level datasets we want to process
}
