% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-region-updates.R
\name{rru_log_outcome}
\alias{rru_log_outcome}
\title{rru_log_outcome
Processes the outcome returned by rru_process_locations to assorted log files
This is handling runtime status information, not actual results values}
\usage{
rru_log_outcome(outcome)
}
\arguments{
\item{outcome}{List produced by rru_process_locations}
}
\description{
rru_log_outcome
Processes the outcome returned by rru_process_locations to assorted log files
This is handling runtime status information, not actual results values
}
