% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-region-updates.R
\name{rru_process_derivatives}
\alias{rru_process_derivatives}
\title{rru_process_derivatives
work out what processing needs to happen for collated derivatives}
\usage{
rru_process_derivatives(derivatives, datasets)
}
\arguments{
\item{derivatives}{List of \code{CollatedDerivative}}

\item{datasets}{List of \code{AbstractDataset}}
}
\description{
rru_process_derivatives
work out what processing needs to happen for collated derivatives
}
