% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-region-updates.R
\name{rru_process_locations}
\alias{rru_process_locations}
\title{rru_process_locations
handles the include / exclude functionality and unpacking the list of datasets.
calls on to update_regional, adding additional logging with improved context.}
\usage{
rru_process_locations(datasets, args, excludes, includes)
}
\arguments{
\item{datasets}{List of AbstractDataset}

\item{args}{List of arguments from cli tool}

\item{excludes}{List of strings, processed from the args}

\item{includes}{List of strings, processed from the args}
}
\value{
List of results for each dataset
}
\description{
rru_process_locations
handles the include / exclude functionality and unpacking the list of datasets.
calls on to update_regional, adding additional logging with improved context.
}
