% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-region-updates.R
\name{run_regional_updates}
\alias{run_regional_updates}
\title{Run Regional Updates
Main function for process - this is probably what you want to call if you are loading a custom
dataset or modified args
Sequences the high level blocks of the process}
\usage{
run_regional_updates(datasets, derivatives, args)
}
\arguments{
\item{datasets}{List of CollatedDerivative  - typically from collated-derivative-list.R}

\item{args}{List of arguments returned by the cli interface (}
}
\description{
Run Regional Updates
Main function for process - this is probably what you want to call if you are loading a custom
dataset or modified args
Sequences the high level blocks of the process
}
