% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-republish.R
\name{run_republish}
\alias{run_republish}
\title{Run republish
Endpoint to recall the publication of the local results for a particular dataset}
\usage{
run_republish(datasets, args)
}
\arguments{
\item{datasets}{List of AbstractDataset - typically from dataset-list.R}

\item{args}{List of arguments returned by the cli interface (}
}
\description{
Run republish
Endpoint to recall the publication of the local results for a particular dataset
}
