% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-resummarise.R
\name{run_resummarise}
\alias{run_resummarise}
\title{Run resummarise
Endpoint to resummarise based on the local results for a particular dataset}
\usage{
run_resummarise(datasets, args)
}
\arguments{
\item{datasets}{List of AbstractDataset - typically from dataset-list.R}

\item{args}{List of arguments returned by the cli interface (}
}
\description{
Run resummarise
Endpoint to resummarise based on the local results for a particular dataset
}
