% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_future}
\alias{setup_future}
\title{Set up parallel processing on all available cores}
\usage{
setup_future(jobs, min_cores_per_worker = 4)
}
\description{
Set up parallel processing on all available cores
}
