% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-regional.R
\name{update_regional}
\alias{update_regional}
\title{Update Regional}
\usage{
update_regional(
  location,
  excludes,
  includes,
  force,
  max_execution_time,
  refresh
)
}
\arguments{
\item{location}{Location object containing information about region}

\item{excludes}{Dataframe containing regions to exclude}

\item{includes}{Dataframe containing the only regions to include}

\item{max_execution_time}{Integer specifying the timeout in seconds}
}
\description{
Processes regional data in an abstract fashion to reduce code duplication
}
